♻️ [Refactor]: shared 계층에서 상위 계층 의존성 제거
