+++ src/shared/components/atoms/Editor.tsx
@@ -1,7 +1,5 @@
 import { useEffect, useRef } from "react";
 import { useFormContext } from "react-hook-form";
-
-import { uploadImage } from "@/entities/study/studies/api/studies.api";
 import MDEditor, {
   commands,
   TextAreaTextApi,
@@ -11,9 +9,10 @@ import rehypeSanitize from "rehype-sanitize";
 
 interface EditorProps {
   name: string;
+  uploadImage: (file: File) => Promise<{ url: string }>;
 }
 
-const Editor = ({ name }: EditorProps) => {
+const Editor = ({ name, uploadImage }: EditorProps) => {
   const { watch, setValue } = useFormContext();
   const value = watch(name);
   const editorRef = useRef<HTMLDivElement>(null);
@@ -21,7 +20,7 @@ const Editor = ({ name }: EditorProps) => {
   // 이미지 업로드 후 마크다운 삽입
   const uploadImageAndInsert = async (file: File) => {
     try {
-      const url = await uploadImage({ file });
+      const url = await uploadImage(file);
       const insert = `![image](${url})`;
       const newValue = (value ?? "") + "\n" + insert;
       setValue(name, newValue);
@@ -47,7 +46,7 @@ const Editor = ({ name }: EditorProps) => {
       if (!file) return;
 
       try {
-        const url = await uploadImage({ file });
+        const url = await uploadImage(file);
         const insert = `![image](${url})`;
 
         api.replaceSelection(insert);
